changeed driectory structure, moved scaling and appearance mode functionality to super classes
