--- customtkinter/widgets/__init__.py
@@ -1,3 +1,0 @@
-from customtkinter.widgets.core_rendering.ctk_canvas import CTkCanvas
-
-CTkCanvas.init_font_character_mapping()
