+++ customtkinter/windows/widgets/__init__.py
@@ -1,0 +1,3 @@
+from customtkinter.windows.widgets.core_rendering.ctk_canvas import CTkCanvas
+
+CTkCanvas.init_font_character_mapping()
