+++ customtkinter/windows/widgets/core_rendering/ctk_canvas.py
