+++ customtkinter/windows/widgets/ctk_entry.py
@@ -2,7 +2,7 @@ import tkinter
 from typing import Union, Tuple
 
 from .core_rendering.ctk_canvas import CTkCanvas
-from ..theme_manager import ThemeManager
+from .theme.theme_manager import ThemeManager
 from .core_rendering.draw_engine import DrawEngine
 from .core_widget_classes.widget_base_class import CTkBaseClass
 from .font.ctk_font import CTkFont
