+++ customtkinter/windows/widgets/ctk_frame.py
@@ -1,7 +1,7 @@
 from typing import Union, Tuple, List
 
 from .core_rendering.ctk_canvas import CTkCanvas
-from ..theme_manager import ThemeManager
+from .theme.theme_manager import ThemeManager
 from .core_rendering.draw_engine import DrawEngine
 from .core_widget_classes.widget_base_class import CTkBaseClass
