+++ customtkinter/windows/widgets/ctk_optionmenu.py
@@ -3,7 +3,7 @@ import sys
 from typing import Union, Tuple, Callable
 
 from .core_rendering.ctk_canvas import CTkCanvas
-from ..theme_manager import ThemeManager
+from .theme.theme_manager import ThemeManager
 from .core_rendering.draw_engine import DrawEngine
 from .core_widget_classes.widget_base_class import CTkBaseClass
 from .core_widget_classes.dropdown_menu import DropdownMenu
