+++ customtkinter/windows/widgets/ctk_tabview.py
@@ -1,6 +1,6 @@
 from typing import Union, Tuple, Dict, List, Callable
 
-from ..theme_manager import ThemeManager
+from .theme.theme_manager import ThemeManager
 from .ctk_frame import CTkFrame
 from .core_widget_classes.widget_base_class import CTkBaseClass
 from .ctk_segmented_button import CTkSegmentedButton
