+++ requirements.txt
@@ -1,1 +1,2 @@
 darkdetect~=0.3.1
+typing-extensions~=4.4.0
